+++ examples/docs-sdk-quickstart/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @executor-js/example-docs-sdk-quickstart
 
+## 0.0.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/plugin-openapi@1.6.6
+  - @executor-js/sdk@1.6.6
+
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/sdk@1.6.5
+  - @executor-js/plugin-openapi@1.6.5
+
 ## 0.0.50
 
 ### Patch Changes