+++ packages/kernel/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @executor-js/codemode-core
 
+## 1.6.6
+
+## 1.6.5
+
 ## 1.6.4
 
 ## 1.6.3