+++ packages/kernel/runtime-quickjs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @executor-js/runtime-quickjs
 
+## 1.6.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/codemode-core@1.6.6
+
+## 1.6.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/codemode-core@1.6.5
+
 ## 1.6.4
 
 ### Patch Changes