+++ packages/plugins/toolkits/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @executor-js/plugin-toolkits
 
+## 1.5.41
+
+### Patch Changes
+
+- Updated dependencies [[`9a1fbd5`](https://github.com/UsefulSoftwareCo/executor/commit/9a1fbd5f0de25f622f303c76f998443c1bb72063)]:
+  - @executor-js/react@1.4.69
+  - @executor-js/api@1.4.69
+  - @executor-js/sdk@1.6.6
+
+## 1.5.40
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @executor-js/sdk@1.6.5
+  - @executor-js/api@1.4.68
+  - @executor-js/react@1.4.68
+
 ## 1.5.39
 
 ### Patch Changes